2024-09-01 update : fixing model alignment prob.

+++ src/uciwweihr_fit.jl
@@ -29,9 +29,9 @@ function uciwweihr_fit(
     params::uciwweihr_model_params
     )
     println("Using uciwweihr_model with wastewater!!!")
-    obstimes_hosp = convert(Vector{Float64}, obstimes_hosp)
-    obstimes_wastewater = convert(Vector{Float64}, obstimes_wastewater)
-    param_change_times = convert(Vector{Float64}, param_change_times)
+    obstimes_hosp = convert(Vector{Int64}, obstimes_hosp)
+    obstimes_wastewater = convert(Vector{Int64}, obstimes_wastewater)
+    param_change_times = convert(Vector{Int64}, param_change_times)
 
 
     my_model = uciwweihr_model(
@@ -63,8 +63,8 @@ function uciwweihr_fit(
     params::uciwweihr_model_params
     )
     println("Using uciwweihr_model without wastewater!!!")
-    obstimes_hosp = convert(Vector{Float64}, obstimes_hosp)
-    param_change_times = convert(Vector{Float64}, param_change_times)
+    obstimes_hosp = convert(Vector{Int64}, obstimes_hosp)
+    param_change_times = convert(Vector{Int64}, param_change_times)
 
 
     my_model = uciwweihr_model(
